Created the createWord function
This function takes a list of Sounds and combines them using the data stored in the sound including the sound duration, the transition period and the sound frame data.
The function will then interpolate between each sound using the data provided to create the word,
Finally an array of 5ms data frames will be outputted to be stored to create the word sound.